backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-dotenv>=1.0.0
google-generativeai>=0.7.2
google-cloud-aiplatform>=1.55.0
httpx>=0.27.0
pydantic>=2.7.1
numpy>=1.26.0
scipy>=1.13.0
